Move PXR wizards into main package.
