+++ modules/praxislive.pxr/src/main/java/org/praxislive/ide/pxr/PXGExportWizardPanel1.java
@@ -19,7 +19,7 @@
  * Please visit https://www.praxislive.org if you need additional information or
  * have any questions.
  */
-package org.praxislive.ide.pxr.wizard;
+package org.praxislive.ide.pxr;
 
 import java.io.File;
 import javax.swing.event.ChangeListener;
@@ -27,7 +27,6 @@ import org.openide.WizardDescriptor;
 import org.openide.util.ChangeSupport;
 import org.openide.util.HelpCtx;
 import org.praxislive.core.ComponentType;
-import org.praxislive.core.ValueFormatException;
 
 class PXGExportWizardPanel1 implements WizardDescriptor.Panel<WizardDescriptor> {
