+++ modules/praxislive.pxr/src/main/resources/org/praxislive/ide/pxr/Bundle.properties
@@ -5,3 +5,35 @@ LBL_PXR_loader_name=PXR Files
 DelegatingArgumentCustomEditor.chooserLabel.text=Choose Editor
 BoundCodeResetPanel.lastSavedOption.text=Last saved
 BoundCodeResetPanel.componentDefaultOption.text=Component default
+PXRVisualPanel1.locationLabel.text=Location
+PXRVisualPanel1.idLabel.text=Root ID
+PXRVisualPanel1.fileLabel.text=File
+PXRVisualPanel1.typeLabel.text=Root Type
+PXRVisualPanel1.idField.text=
+PXRVisualPanel1.browseButton.text=Browse ...
+PXGExportVisualPanel1.nameLabel.text=Name
+PXGExportVisualPanel1.locationLabel.text=Location
+PXGExportVisualPanel1.fileLabel.text=File
+PXGExportVisualPanel1.nameField.text=
+PXGExportVisualPanel1.locationField.text=
+PXGExportVisualPanel1.fileField.text=
+PXGExportVisualPanel1.browseButton.text=Browse ...
+PXGExportVisualPanel1.paletteCheckbox.text=Add to palette
+PXGExportVisualPanel1.paletteCategoryLabel.text=Palette category
+PXGExportVisualPanel1.paletteCategoryField.text=
+PXRVisualPanel1.idField.text_1=
+PXGExportVisualPanel1.paletteCheckbox.text_1=Add to palette
+PXGExportVisualPanel1.paletteCategoryField.text_1=
+PXGExportVisualPanel1.paletteCategoryLabel.text_1=Palette category
+PXGExportVisualPanel1.fileField.text_1=
+PXGExportVisualPanel1.browseButton.text_1=Browse ...
+PXGExportVisualPanel1.locationField.text_1=
+PXGExportVisualPanel1.nameField.text_1=
+PXGExportVisualPanel1.fileLabel.text_1=File
+PXGExportVisualPanel1.locationLabel.text_1=Location
+PXGExportVisualPanel1.nameLabel.text_1=Name
+PXRVisualPanel1.typeLabel.text_1=Root Type
+PXRVisualPanel1.fileLabel.text_1=File
+PXRVisualPanel1.browseButton.text_1=Browse ...
+PXRVisualPanel1.locationLabel.text_1=Location
+PXRVisualPanel1.idLabel.text_1=Root ID
