--- modules/praxislive.pxr/src/main/resources/org/praxislive/ide/pxr/wizard/Bundle.properties
@@ -1,16 +1,0 @@
-PXRVisualPanel1.locationLabel.text=Location
-PXRVisualPanel1.idLabel.text=Root ID
-PXRVisualPanel1.fileLabel.text=File
-PXRVisualPanel1.typeLabel.text=Root Type
-PXRVisualPanel1.idField.text=
-PXRVisualPanel1.browseButton.text=Browse ...
-PXGExportVisualPanel1.nameLabel.text=Name
-PXGExportVisualPanel1.locationLabel.text=Location
-PXGExportVisualPanel1.fileLabel.text=File
-PXGExportVisualPanel1.nameField.text=
-PXGExportVisualPanel1.locationField.text=
-PXGExportVisualPanel1.fileField.text=
-PXGExportVisualPanel1.browseButton.text=Browse ...
-PXGExportVisualPanel1.paletteCheckbox.text=Add to palette
-PXGExportVisualPanel1.paletteCategoryLabel.text=Palette category
-PXGExportVisualPanel1.paletteCategoryField.text=
